api/requirements.txt
aiomysql==0.1.1
annotated-types==0.7.0
anyio==4.7.0
Authlib==1.4.0
bcrypt==4.2.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
dynaconf==3.2.6
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-sa==0.2.0
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.21.1
watchfiles==1.0.3
websockets==14.1
